+++ Classes/URKArchive.h
@@ -98,6 +98,11 @@ typedef NS_ENUM(NSInteger, URKErrorCode) {
      *  User cancelled the operation
      */
     URKErrorCodeUserCancelled = 102,
+    
+    /**
+     *  Error converting string to UTF-8
+     */
+    URKErrorCodeStringConversion = 103,
 };
 
 typedef NSString *const URKProgressInfoKey;
@@ -283,6 +288,24 @@ extern NSString *URKErrorDomain;
  *  @return Returns a list of URKFileInfo objects, which contain metadata about the archive's files, or nil if an error was encountered
  */
 - (nullable NSArray<URKFileInfo*> *)listFileInfo:(NSError **)error;
+
+/**
+ *  Iterates the header of the archive, calling the block with each archived file's info.
+ *
+ *  WARNING: There is no filtering of duplicate header entries. If a file is listed twice, `action`
+ *  will be called twice with that file's path
+ *
+ *  @param action The action to perform using the data. Must be non-nil
+ *
+ *       - *fileInfo* The metadata of the file within the archive
+ *       - *stop*     Set to YES to stop reading the archive
+ *
+ *  @param error Contains an NSError object when there was an error reading the archive
+ *
+ *  @return Returns NO if an error was encountered
+ */
+- (BOOL) iterateFileInfo:(void(^)(URKFileInfo *fileInfo, BOOL *stop))action
+                   error:(NSError **)error;
 
 /**
  *  Lists the URLs of volumes in a single- or multi-volume archive