+++ Classes/URKArchive.mm
@@ -44,6 +44,7 @@ NS_DESIGNATED_INITIALIZER
 
 @property (strong) NSData *fileBookmark;
 @property (strong) BOOL(^bufferedReadBlock)(NSData *dataChunk);
+@property (strong) BOOL(^shouldCancelBlock)();
 
 @property (strong) NSObject *threadLock;
 
@@ -374,30 +375,58 @@ NS_DESIGNATED_INITIALIZER
 {
     URKCreateActivity("Listing File Info");
 
-    NSArray<URKFileInfo*> *allFileInfo = [self allFileInfo:error];
+    NSMutableSet<NSString*> *distinctFilenames = [NSMutableSet set];
+    NSMutableArray<URKFileInfo*> *distinctFileInfo = [NSMutableArray array];
+    NSError *innerError = nil;
+
+    BOOL wasSuccessful = [self iterateFileInfo:^(URKFileInfo * _Nonnull fileInfo, BOOL * _Nonnull stop) {
+        if (![distinctFilenames containsObject:fileInfo.filename]) {
+            [distinctFileInfo addObject:fileInfo];
+            [distinctFilenames addObject:fileInfo.filename];
+        } else {
+            URKLogDebug("Skipping %{public}@ from list of file info, since it's already represented (probably from another archive volume)", fileInfo.filename);
+        }
+    }
+                    error:&innerError];
     
-    if (!allFileInfo) {
-        URKLogError("-allFileInfo: returned an error")
+    if (!wasSuccessful) {
+        URKLogError("Failed to iterate file info: %{public}@", innerError);
+        
+        if (error && innerError) {
+            *error = innerError;
+        }
+        
         return nil;
     }
     
-    URKLogDebug("Found %lu total file info items", (unsigned long)allFileInfo.count);
+    URKLogDebug("Found %lu file info items", (unsigned long)distinctFileInfo.count);
+    return [NSArray arrayWithArray:distinctFileInfo];
+}
+
+- (BOOL) iterateFileInfo:(void(^)(URKFileInfo *fileInfo, BOOL *stop))action
+                   error:(NSError * __autoreleasing *)error
+{
+    URKCreateActivity("Iterating File Info");
+    NSAssert(action != nil, @"'action' is a required argument");
+
+    NSError *innerError = nil;
+
+    URKLogDebug("Beginning to iterate through contents of %{public}@", self.filename);
     
-    NSMutableSet<NSString*> *distinctFilenames = [NSMutableSet set];
-    NSMutableArray<URKFileInfo*> *distinctFileInfo = [NSMutableArray array];
+    BOOL wasSuccessful = [self iterateAllFileInfo:action
+                                            error:&innerError];
     
-    for (URKFileInfo *info in allFileInfo) {
-        if (![distinctFilenames containsObject:info.filename]) {
-            [distinctFileInfo addObject:info];
-            [distinctFilenames addObject:info.filename];
-        } else {
-            URKLogDebug("Skipping %{public}@ from list of file info, since it's already represented (probably from another archive volume)", info.filename);
+    if (!wasSuccessful) {
+        URKLogError("Failed to iterate all file info: %{public}@", innerError);
+        
+        if (error && innerError) {
+            *error = innerError;
         }
+        
+        return NO;
     }
     
-    URKLogDebug("Found %lu distinct file info items", (unsigned long)distinctFileInfo.count);
-
-    return [NSArray arrayWithArray:distinctFileInfo];
+    return YES;
 }
 
 - (nullable NSArray<NSURL*> *)listVolumeURLs:(NSError * __autoreleasing *)error
@@ -504,10 +533,34 @@ NS_DESIGNATED_INITIALIZER
                 return;
             }
 
-            if ((PFCode = RARProcessFile(welf.rarFile, RAR_EXTRACT, (char *) filePath.UTF8String, NULL)) != 0) {
+            char cFilePath[2048];
+            BOOL utf8ConversionSucceeded = [filePath getCString:cFilePath
+                                                      maxLength:sizeof(cFilePath)
+                                                       encoding:NSUTF8StringEncoding];
+            if (!utf8ConversionSucceeded) {
                 NSString *errorName = nil;
-                [self assignError:innerError code:(NSInteger)PFCode errorName:&errorName];
-                URKLogError("Error extracting file: %{public}@ (%d)", errorName, PFCode);
+                [self assignError:innerError code:URKErrorCodeStringConversion errorName:&errorName];
+                URKLogError("Error converting file to UTF-8 (buffer too short?)");
+                result = NO;
+                return;
+            }
+            
+            RARSetCallback(welf.rarFile, AllowCancellationCallbackProc, (long)(__bridge void *) self);
+            self.shouldCancelBlock = ^BOOL{
+                URKCreateActivity("shouldCancelBlock");
+                URKLogDebug("Progress.isCancelled: %{public}@", progress.isCancelled ? @"YES" : @"NO")
+                return progress.isCancelled;
+            };
+
+            if ((PFCode = RARProcessFile(welf.rarFile, RAR_EXTRACT, cFilePath, NULL)) != 0) {
+                RARSetCallback(welf.rarFile, NULL, NULL);
+                self.shouldCancelBlock = nil;
+
+                NSString *errorName = nil;
+                
+                NSInteger errorCode = progress.isCancelled ? URKErrorCodeUserCancelled : PFCode;
+                [self assignError:innerError code:errorCode errorName:&errorName];
+                URKLogError("Error extracting file: %{public}@ (%ld)", errorName, (long)errorCode);
                 result = NO;
                 return;
             }
@@ -531,6 +584,9 @@ NS_DESIGNATED_INITIALIZER
 
             bytesDecompressed += fileInfo.uncompressedSize;
         }
+        
+        RARSetCallback(welf.rarFile, NULL, NULL);
+        self.shouldCancelBlock = nil;
 
         if (RHCode != ERAR_SUCCESS && RHCode != ERAR_END_ARCHIVE) {
             NSString *errorName = nil;
@@ -659,6 +715,9 @@ NS_DESIGNATED_INITIALIZER
         
         URKLogInfo("Processing file...");
         PFCode = RARProcessFile(welf.rarFile, RAR_TEST, NULL, NULL);
+        
+        RARSetCallback(welf.rarFile, NULL, NULL);
+        self.bufferedReadBlock = nil;
         
         if (progress.isCancelled) {
             NSString *errorName = nil;
@@ -830,7 +889,7 @@ NS_DESIGNATED_INITIALIZER
 {
     URKCreateActivity("Extracting Buffered Data");
 
-    NSError *innerError = nil;
+    NSError *actionError = nil;
 
     NSProgress *progress = [self beginProgressOperation:0];
 
@@ -907,6 +966,9 @@ NS_DESIGNATED_INITIALIZER
 
         URKLogDebug("Processing file...");
         PFCode = RARProcessFile(welf.rarFile, RAR_TEST, NULL, NULL);
+        
+        RARSetCallback(welf.rarFile, NULL, NULL);
+        self.bufferedReadBlock = nil;
 
         if (progress.isCancelled) {
             NSString *errorName = nil;
@@ -920,17 +982,17 @@ NS_DESIGNATED_INITIALIZER
             [self assignError:innerError code:(NSInteger)PFCode errorName:&errorName];
             URKLogError("Error processing file: %{public}@ (%d)", errorName, PFCode);
         }
-    } inMode:RAR_OM_EXTRACT error:&innerError];
+    } inMode:RAR_OM_EXTRACT error:&actionError];
 
     if (error) {
-        *error = innerError ? innerError : nil;
+        *error = actionError;
 
-        if (innerError) {
-            URKLogError("Error reading buffered data from file\nfilePath: %{public}@\nerror: %{public}@", filePath, innerError);
+        if (actionError) {
+            URKLogError("Error reading buffered data from file\nfilePath: %{public}@\nerror: %{public}@", filePath, actionError);
         }
     }
 
-    return success && !innerError;
+    return success && !actionError;
 }
 
 - (BOOL)isPasswordProtected
@@ -1118,6 +1180,23 @@ int CALLBACK BufferedReadCallbackProc(UINT msg, long UserData, long P1, long P2)
     return 0;
 }
 
+int CALLBACK AllowCancellationCallbackProc(UINT msg, long UserData, long P1, long P2) {
+    URKCreateActivity("AllowCancellationCallbackProc");
+    URKArchive *refToSelf = (__bridge URKArchive *)(void *)UserData;
+
+    if (![refToSelf shouldCancelBlock]) {
+        return 0;
+    }
+    
+    BOOL shouldCancel = refToSelf.shouldCancelBlock();
+    if (shouldCancel) {
+        URKLogDebug("Operation cancelled in shouldCancelBlock()");
+        refToSelf.shouldCancelBlock = nil;
+    }
+    
+    return shouldCancel ? -1 : 0;
+}
+
 
 
 #pragma mark - Private Methods
@@ -1212,8 +1291,19 @@ int CALLBACK BufferedReadCallbackProc(UINT msg, long UserData, long P1, long P2)
 
     if(aPassword != nil) {
         URKLogDebug("Setting password...");
-        char *password = (char *) [aPassword UTF8String];
-        RARSetPassword(self.rarFile, password);
+        
+        char cPassword[2048];
+        BOOL utf8ConversionSucceeded = [aPassword getCString:cPassword
+                                                   maxLength:sizeof(cPassword)
+                                                    encoding:NSUTF8StringEncoding];
+        if (!utf8ConversionSucceeded) {
+            NSString *errorName = nil;
+            [self assignError:error code:URKErrorCodeStringConversion errorName:&errorName];
+            URKLogError("Error converting password to UTF-8 (buffer too short?)");
+            return NO;
+        }
+        
+        RARSetPassword(self.rarFile, cPassword);
     }
 
 	return YES;
@@ -1239,13 +1329,15 @@ int CALLBACK BufferedReadCallbackProc(UINT msg, long UserData, long P1, long P2)
     return YES;
 }
 
-- (NSArray<URKFileInfo *> *) allFileInfo:(NSError * __autoreleasing *)error {
+- (BOOL) iterateAllFileInfo:(void(^)(URKFileInfo *fileInfo, BOOL *stop))action
+                      error:(NSError * __autoreleasing *)error
+{
     URKCreateActivity("-allFileInfo:");
+    NSAssert(action != nil, @"'action' is a required argument");
     
-    __block NSMutableArray *fileInfos = [NSMutableArray array];
     __weak URKArchive *welf = self;
     
-    BOOL success = [self performActionWithArchiveOpen:^(NSError **innerError) {
+    BOOL wasSuccessful = [self performActionWithArchiveOpen:^(NSError **innerError) {
         URKCreateActivity("Performing List Action");
         
         int RHCode = 0, PFCode = 0;
@@ -1253,15 +1345,21 @@ int CALLBACK BufferedReadCallbackProc(UINT msg, long UserData, long P1, long P2)
         URKLogDebug("Reading through RAR header looking for files...");
         
         while ((RHCode = RARReadHeaderEx(welf.rarFile, welf.header)) == 0) {
-            URKLogDebug("Adding object");
-            [fileInfos addObject:[URKFileInfo fileInfo:welf.header]];
+            URKLogDebug("Calling iterateAllFileInfo handler");
+            BOOL shouldStop = NO;
+            URKFileInfo *info = [URKFileInfo fileInfo:welf.header];
+            action(info, &shouldStop);
+            
+            if (shouldStop) {
+                URKLogDebug("iterateAllFileInfo got signal to stop");
+                return;
+            }
             
             URKLogDebug("Skipping to next file...");
             if ((PFCode = RARProcessFile(welf.rarFile, RAR_SKIP, NULL, NULL)) != 0) {
                 NSString *errorName = nil;
                 [self assignError:innerError code:(NSInteger)PFCode errorName:&errorName];
                 URKLogError("Error skipping to next header file: %{public}@ (%d)", errorName, PFCode);
-                fileInfos = nil;
                 return;
             }
         }
@@ -1270,11 +1368,31 @@ int CALLBACK BufferedReadCallbackProc(UINT msg, long UserData, long P1, long P2)
             NSString *errorName = nil;
             [self assignError:innerError code:RHCode errorName:&errorName];
             URKLogError("Error reading RAR header: %{public}@ (%d)", errorName, RHCode);
-            fileInfos = nil;
         }
     } inMode:RAR_OM_LIST_INCSPLIT error:error];
     
-    if (!success || !fileInfos) {
+    return wasSuccessful;
+}
+
+- (NSArray<URKFileInfo *> *) allFileInfo:(NSError * __autoreleasing *)error {
+    URKCreateActivity("-allFileInfo:");
+    
+    NSMutableArray *fileInfos = [NSMutableArray array];
+    NSError *innerError = nil;
+    
+    URKLogDebug("Iterating all file info");
+    BOOL wasSuccessful = [self iterateAllFileInfo:^(URKFileInfo *fileInfo, BOOL *stop) {
+        [fileInfos addObject:fileInfo];
+    }
+                       error:&innerError];
+    
+    if (!wasSuccessful || !fileInfos) {
+        URKLogError("File info iteration was not successful: %{public}@", innerError);
+
+        if (error && innerError) {
+            *error = innerError;
+        }
+        
         return nil;
     }
     
@@ -1362,6 +1480,12 @@ int CALLBACK BufferedReadCallbackProc(UINT msg, long UserData, long P1, long P2)
         case URKErrorCodeUserCancelled:
             errorName = @"ERAR_USER_CANCELLED";
             detail = NSLocalizedStringFromTableInBundle(@"User cancelled the operation in progress", @"UnrarKit", _resources, @"Error detail string");
+            break;
+            
+
+        case URKErrorCodeStringConversion:
+            errorName = @"ERAR_UTF8_PATH_CONVERSION";
+            detail = NSLocalizedStringFromTableInBundle(@"Error converting a string to UTF-8", @"UnrarKit", _resources, @"Error detail string");
             break;
 
         default: