+++ Scripts/push-output.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-echo
+set -ev
 
 # Only potentially push to CocoaPods when it's a tagged build
 if [ -z "$TRAVIS_TAG" ]; then
@@ -16,8 +16,11 @@ else
     echo -e "\nTag looks like a version number: $TRAVIS_TAG"
 fi
 
+# Skip tests because they're assumed to have passed during the cocoapod-validate script or else
+# this script wouldn't run. Allow warnings, because prior validation doesn't (and it patches)
+# around the one warning in the UnRAR library I can't silence
 echo -e "\nLinting podspec..."
-pod spec lint --fail-fast
+pod spec lint --fail-fast --skip-tests --allow-warnings
 
 if [ $? -ne 0 ]; then
     echo -e "\nPodspec failed lint. Run again with --verbose to troubleshoot"
@@ -28,8 +31,9 @@ echo -e "\nExporting Carthage archive...\n"
 # Exports ARCHIVE_PATH, used below
 source ./Scripts/archive-carthage.sh
 
+# Skip tests and allow warnings for reasons stated above
 echo -e "\nPushing to CocoaPods...\n"
-pod trunk push
+pod trunk push --skip-tests --allow-warnings
 
 # If push is successful, add release to GitHub
 if [ $? -ne 0 ]; then